Strip down the interface and add matrix fancy
I decided that georef should be a lighter package, that it shouldn't try
to handle georeferencing weird formats on its own. Rather, it should
contain *all* of the georeferencing math, and allow georeferencing of
the most basic cases. It should also expose an api so it can be used by
a downstream that has to do more fancy work (e.g. an airborne MTA
system).